Currency: Use BigDecimal to store exchange rates
Particularly the inverse exchange rates need a better precision so that
converting a currency amount back and forth does not introduce rounding
errors.

+++ name.abuchen.portfolio.tests/src/name/abuchen/portfolio/money/impl/ExchangeRateTimeSeriesImplTest.java
@@ -2,8 +2,10 @@ package name.abuchen.portfolio.money.impl;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+
+import java.math.BigDecimal;
+
 import name.abuchen.portfolio.money.ExchangeRate;
-import name.abuchen.portfolio.money.impl.ExchangeRateTimeSeriesImpl;
 import name.abuchen.portfolio.util.Dates;
 
 import org.junit.Test;
@@ -15,23 +17,23 @@ public class ExchangeRateTimeSeriesImplTest
     public void testLookupOfExchangeRate()
     {
         ExchangeRateTimeSeriesImpl series = new ExchangeRateTimeSeriesImpl();
-        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), 1));
-        series.addRate(new ExchangeRate(Dates.date("2014-12-02"), 2));
-        series.addRate(new ExchangeRate(Dates.date("2014-12-03"), 3));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(1)));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-02"), BigDecimal.valueOf(2)));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-03"), BigDecimal.valueOf(3)));
 
-        assertThat(series.lookupRate(Dates.date("2014-11-30")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-02")).get().getValue(), is(2L));
-        assertThat(series.lookupRate(Dates.date("2014-12-03")).get().getValue(), is(3L));
-        assertThat(series.lookupRate(Dates.date("2014-12-04")).get().getValue(), is(3L));
+        assertThat(series.lookupRate(Dates.date("2014-11-30")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-02")).get().getValue(), is(BigDecimal.valueOf(2)));
+        assertThat(series.lookupRate(Dates.date("2014-12-03")).get().getValue(), is(BigDecimal.valueOf(3)));
+        assertThat(series.lookupRate(Dates.date("2014-12-04")).get().getValue(), is(BigDecimal.valueOf(3)));
     }
 
     @Test
     public void testAddingOfExchangeRates()
     {
         ExchangeRateTimeSeriesImpl series = new ExchangeRateTimeSeriesImpl(null, "EUR", "USD");
-        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), 1));
-        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), 2));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(1)));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(2)));
 
         assertProperties(series);
     }
@@ -40,8 +42,8 @@ public class ExchangeRateTimeSeriesImplTest
     public void testCreationFromTemplate()
     {
         ExchangeRateTimeSeriesImpl template = new ExchangeRateTimeSeriesImpl(null, "EUR", "USD");
-        template.addRate(new ExchangeRate(Dates.date("2014-12-01"), 1));
-        template.addRate(new ExchangeRate(Dates.date("2014-12-01"), 2));
+        template.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(1)));
+        template.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(2)));
 
         ExchangeRateTimeSeriesImpl series = new ExchangeRateTimeSeriesImpl(template);
 
@@ -52,22 +54,22 @@ public class ExchangeRateTimeSeriesImplTest
     {
         assertThat(series.getBaseCurrency(), is("EUR"));
         assertThat(series.getTermCurrency(), is("USD"));
-        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(2L));
-        assertThat(series.getLatest().get(), is(new ExchangeRate(Dates.date("2014-12-01"), 2)));
+        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(BigDecimal.valueOf(2)));
+        assertThat(series.getLatest().get(), is(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(2))));
     }
 
     @Test
     public void testLookupOfExchangeRateWithGaps()
     {
         ExchangeRateTimeSeriesImpl series = new ExchangeRateTimeSeriesImpl();
-        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), 1));
-        series.addRate(new ExchangeRate(Dates.date("2014-12-03"), 3));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(1)));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-03"), BigDecimal.valueOf(3)));
 
-        assertThat(series.lookupRate(Dates.date("2014-11-30")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-02")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-03")).get().getValue(), is(3L));
-        assertThat(series.lookupRate(Dates.date("2014-12-04")).get().getValue(), is(3L));
+        assertThat(series.lookupRate(Dates.date("2014-11-30")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-02")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-03")).get().getValue(), is(BigDecimal.valueOf(3)));
+        assertThat(series.lookupRate(Dates.date("2014-12-04")).get().getValue(), is(BigDecimal.valueOf(3)));
     }
 
     @Test
@@ -82,11 +84,11 @@ public class ExchangeRateTimeSeriesImplTest
     public void testLookupOfExchangeRateWithOneRateOnly()
     {
         ExchangeRateTimeSeriesImpl series = new ExchangeRateTimeSeriesImpl();
-        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), 1));
+        series.addRate(new ExchangeRate(Dates.date("2014-12-01"), BigDecimal.valueOf(1)));
 
-        assertThat(series.lookupRate(Dates.date("2014-11-30")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(1L));
-        assertThat(series.lookupRate(Dates.date("2014-12-02")).get().getValue(), is(1L));
+        assertThat(series.lookupRate(Dates.date("2014-11-30")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-01")).get().getValue(), is(BigDecimal.valueOf(1)));
+        assertThat(series.lookupRate(Dates.date("2014-12-02")).get().getValue(), is(BigDecimal.valueOf(1)));
     }
 
 }
